pyproject.toml
[project]
name = "discordbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
